pyproject.toml
[project]
name = "cnn"
version = "0.1.0"
description = ""
authors = [
    {name = "Tobias",email = "tobias.bp2004@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow (>=12.1.0,<13.0.0)",
    "numpy (>=2.4.2,<3.0.0)",
    "scikit-image (>=0.26.0,<0.27.0)",
    "scikit-learn (>=1.8.0,<2.0.0)",
    "joblib (>=1.5.3,<2.0.0)",
    "opencv-python (>=4.13.0.92,<5.0.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
